sched/fair: cast se->slice to u64 in set_protect_slice()
This fix the following warning:
../kernel/sched/fair.c:857:10: error: comparison of distinct pointer types ('typeof (slice) *' (aka 'unsigned long long *') and 'typeof (se->slice) *' (aka 'long long *')) [-Werror,-Wcompare-distinct-pointer-types]
  857 |         slice = min(slice, s  AR      sound/arm/built-in.a
e->slice);
      |                 ^~~~~~~~~~~~~~~~~~~~~
../include/linux/kernel.h:874:19: note: expanded from macro 'min'
  874 | #define min(x, y)       __careful_cmp(x, y, <)
      |                         ^~~~~~~~~~~~~~~~~~~~~~
../include/linux/kernel.h:865:24: note: expanded from macro '__careful_cmp'
  865 |         __builtin_choose_expr(__safe_cmp(x, y), \
      |                               ^~~~~~~~~~~~~~~~
../include/linux/kernel.h:855:4: note: expanded from macro '__safe_cmp'
  855 |                 (__typecheck(x, y) && __no_side_effects(x, y))
      |                  ^~~~~~~~~~~~~~~~~
../include/linux/kernel.h:841:29: note: expanded from macro '__typecheck'
  841 |                 (!!(sizeof((typeof(x) *)1 == (typeof(y) *)1)))
      |

+++ kernel/sched/fair.c
@@ -854,7 +854,7 @@ static inline void set_protect_slice(struct cfs_rq *cfs_rq, struct sched_entity 
 	if (sched_feat(RUN_TO_PARITY))
 		slice = cfs_rq_min_slice(cfs_rq);
 
-	slice = min(slice, se->slice);
+	slice = min(slice, (u64)se->slice);
 	if (slice != se->slice)
 		vprot = min_vruntime(vprot, se->vruntime + calc_delta_fair(slice, se));
